alpine

+++ modules/alpine_syscall_hooks/src/hook.c
@@ -224,7 +224,7 @@ int install_hooks(void) {
     syscall_table = lookup_syscall_table();// (long unsigned int*)kallsyms_lookup_name("sys_call_table");
     echo("sys_call_table address %lu\n", syscall_table);
     if(!syscall_table) return 0;
-    make_vm_rw((long unsigned int)syscall_table);
+    make_vm_rw(syscall_table);
     original_syscall_table[__NR_read] = syscall_table[__NR_read];
     syscall_table[__NR_read] = (long unsigned int)alpine_ksys_read;
     make_vm_ro((long unsigned int)syscall_table);
@@ -233,7 +233,7 @@ int install_hooks(void) {
 void uninstall_hooks(void){
     if(!syscall_table) return ;
     echo("sys_call_table address %p\n", syscall_table);
-    make_vm_rw((long unsigned int)syscall_table);
+    make_vm_rw(syscall_table);
     syscall_table[__NR_read] = original_syscall_table[__NR_read];
     make_vm_ro((long unsigned int)syscall_table);
 }
